@@ -4,7 +4,7 @@ import android.app.Activity;
 import android.content.Intent;
 import android.graphics.Color;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.SurfaceView;
 import android.view.View;
 
@@ -13,19 +13,10 @@ import com.yt.android.zxing.callback.OnDecodedResultListener;
 import com.yt.android.zxing.camera.PreviewHelper;
 import com.yt.android.zxing.view.ViewfinderView;
 
-/**
- * @author BayMax?Yi
- * @time 2019/6/11 13:41
- * @modiffy
- * @modiffyTime 2019/6/11 13:41
- * @describe
- */
 public class CaptureActivity extends Activity {
     private SurfaceView surfaceView;
     private ViewfinderView viewfinderView;
     private PreviewHelper previewHelper;
-
-    private final String TAG = "MainActivity";
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -55,18 +46,19 @@ public class CaptureActivity extends Activity {
 
     private void configViewFiderView(ViewfinderView viewfinderView) {
         viewfinderView.setAnimatorDuretion(1500);
-        viewfinderView.setConorColor(Color.GREEN);
-        viewfinderView.setCornorLength(60);
-        viewfinderView.setLineColors(new int[]{Color.argb(10, 155, 255, 60),
-                Color.argb(30, 155, 255, 60),
-                Color.argb(50, 155, 255, 60),
-                Color.argb(70, 155, 255, 60),
-                Color.argb(90, 155, 255, 60),
-                Color.argb(70, 155, 255, 60),
-                Color.argb(50, 155, 255, 60),
-                Color.argb(30, 155, 255, 60),
-                Color.argb(10, 155, 255, 60),
-                Color.argb(5, 155, 255, 60),});
+        viewfinderView.setConorColor(Color.rgb(9, 166, 253));
+        viewfinderView.setCornorLength(70);
+        viewfinderView.setLineColors(new int[]{
+                Color.argb(25, 9, 166, 253),
+                Color.argb(76, 9, 166, 253),
+                Color.argb(127, 9, 166, 253),
+                Color.argb(178, 9, 166, 253),
+                Color.argb(255, 9, 166, 253),
+                Color.argb(178, 9, 166, 253),
+                Color.argb(127, 9, 166, 253),
+                Color.argb(76, 9, 166, 253),
+                Color.argb(25, 9, 166, 253),
+                Color.argb(5, 9, 166, 253),});
         viewfinderView.setLinePercents(new float[]{0.1f, 0.2f, 0.3f, 0.4f, 0.5f, 0.6f, 0.7f, 0.8f, 0.9f, 1f});
     }
 
