@@ -3,7 +3,7 @@ package com.baymax.baymaxscanner;
 import android.app.Activity;
 import android.graphics.Bitmap;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.TypedValue;
 import android.view.View;
 import android.widget.ImageView;
